Working Java LSP and lsp-zero
The LSP configuration was changed to use the lsp-zero config plugin.
This is intetnded to stream line the process of configuring the LSP
clients. The plugin also standardizes the key bindings to be used in the
context of LSP.

The Java LSP is not managet through the lsp-zero plugin, rather with the
mfussenegger/nvim-jdtls. The LSP server is managed using homebrew.